refactor, fix some layout, proper folder naming.
