+++ lib/main.dart
@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 import 'package:tb_e_health/models/anonymous_user.dart';
 import 'package:tb_e_health/screens/wrapper.dart';
 import 'package:tb_e_health/services/auth_service.dart';
-import 'package:tb_e_health/theme.dart';
+import 'package:tb_e_health/screens/shared/theme.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
