+++ lib/models/faq_data.dart
@@ -1,6 +1,3 @@
-import 'dart:convert';
-
-import 'package:flutter/services.dart';
 
 class FaqData {
   final String title;
