+++ lib/screens/home/home.dart
@@ -1,11 +1,10 @@
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get_navigation/src/root/get_material_app.dart';
-import 'package:tb_e_health/screens/daily_progress_board.dart';
-import 'package:tb_e_health/screens/navigations.dart';
+import 'package:tb_e_health/screens/profile/daily_progress_board.dart';
+import 'package:tb_e_health/screens/navigation_bar/navigations.dart';
 import 'package:tb_e_health/screens/teleconsultation/scheduler_screen.dart';
-import 'package:tb_e_health/screens/teleconsultation/session_screen.dart';
-import 'package:tb_e_health/theme.dart';
+import 'package:tb_e_health/screens/shared/theme.dart';
 
 class Home extends StatelessWidget {
 
@@ -23,7 +22,6 @@ class Home extends StatelessWidget {
           "/home": (context) => Navigations(),
           "/scheduler": (context) => SchedulerScreen(),
           "/dailyProgress": (context) => DailyProgressBoardScreen(),
-          "/teleSession": (context) => SessionScreen(),
         });
 
   }
