+++ lib/screens/navigation_bar/qna.dart
@@ -1,13 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:persistent_bottom_nav_bar/persistent-tab-view.dart';
-import 'package:tb_e_health/models/faq_data.dart';
-import 'package:tb_e_health/screens/chatbot/live_chat.dart';
-import 'package:tb_e_health/screens/navigations.dart';
-import 'package:tb_e_health/screens/tb_info/faq.dart';
-import 'package:tb_e_health/screens/tb_info/tb_dots.dart';
-import 'package:tb_e_health/screens/tb_info/tb_overview.dart';
-import 'package:tb_e_health/screens/tb_info/tb_side_effect_info.dart';
-import 'package:tb_e_health/screens/tb_info/tb_symptom_info.dart';
+import 'package:tb_e_health/screens/useful_info/faq.dart';
+import 'package:tb_e_health/screens/useful_info/tb_overview.dart';
+import 'package:tb_e_health/screens/useful_info/tb_side_effect_info.dart';
+import 'package:tb_e_health/screens/useful_info/tb_symptom_info.dart';
 import 'package:tb_e_health/services/load_data.dart';
 
 class QnA extends StatefulWidget {
