+++ lib/screens/teleconsultation/scheduler_screen.dart
@@ -2,17 +2,13 @@ import 'dart:collection';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:persistent_bottom_nav_bar/persistent-tab-view.dart';
 import 'package:table_calendar/table_calendar.dart';
-import 'package:tb_e_health/Custom%20Widgets/custom_alert_dialog.dart';
+import 'package:tb_e_health/custom_widgets/custom_alert_dialog.dart';
 import 'package:tb_e_health/models/active_user.dart';
 import 'package:tb_e_health/models/appointment.dart';
-import 'package:tb_e_health/screens/chatbot/live_chat.dart';
 import 'package:tb_e_health/screens/shared/common_app_bar.dart';
-import 'package:tb_e_health/screens/teleconsultation/datetime_picker.dart';
 import 'package:tb_e_health/screens/teleconsultation/request_appointment_screen.dart';
-
-import 'package:tb_e_health/utils.dart';
+import 'package:tb_e_health/screens/shared/utils.dart';
 
 class SchedulerScreen extends StatefulWidget {
   @override
@@ -23,16 +19,11 @@ class _SchedulerScreenState extends State<SchedulerScreen> {
   DateTime focusDateTime = DateTime.now();
 
   void _requestAppointment() async {
-    String received = await Navigator.push(
+    String? received = await Navigator.push(
         context, MaterialPageRoute(builder: (_) => RequestApointmentScreen()));
     setState(() {
       // so that the page refresh, and show the newly added record.
     });
-    // pushNewScreen(
-    //   context,
-    //   screen: RequestApointmentScreen(),
-    //   withNavBar: false,
-    // );
   }
 
   @override
@@ -152,11 +143,9 @@ class _SchedulerScreenContentState extends State<SchedulerScreenContent> {
                         icon: Icon(Icons.remove_circle_outlined),
                       ),
                       onTap: () {
-                      print('${value[index]}');
-                      setState(() {
-
-                      });
-                    },
+                        print('${value[index]}');
+                        setState(() {});
+                      },
                       title: Text('${value[index]}'),
                     ),
                   );
