+++ lib/screens/useful_info/tb_side_effect_info.dart
